+++ src/Vaguul.CodexAccountSwitcher/Services/CodexLoginMode.cs
@@ -1,0 +1,7 @@
+namespace Vaguul.CodexAccountSwitcher.Services;
+
+public enum CodexLoginMode
+{
+    Browser,
+    DeviceCode
+}